test(e2e): share runtime message helper
