+++ e2e/alarmSchedulers.spec.ts
@@ -49,6 +49,7 @@ import {
   setPlasmoStorageValue,
 } from "~~/e2e/utils/extensionState"
 import { waitForExtensionRoot } from "~~/e2e/utils/lazyLoading"
+import { sendTypedRuntimeMessageFromPage } from "~~/e2e/utils/runtimeMessaging"
 
 const AUTO_CHECKIN_LEGACY_ALARM_NAME = "autoCheckin"
 const AUTO_CHECKIN_DAILY_ALARM_NAME = "autoCheckinDaily"
@@ -151,26 +152,6 @@ async function openExtensionPage(page: Page, extensionId: string) {
   await page.goto(`chrome-extension://${extensionId}/${OPTIONS_PAGE_PATH}`)
   await waitForExtensionRoot(page)
   await expectPermissionOnboardingHidden(page)
-}
-
-async function sendTypedRuntimeMessageFromPage<TResponse>(
-  page: Page,
-  type: string,
-  data?: Record<string, unknown>,
-): Promise<TResponse> {
-  return await page.evaluate(
-    async ({ type, data }) => {
-      const chromeApi = (globalThis as any).chrome
-      const response = await chromeApi.runtime.sendMessage({
-        id: Date.now(),
-        type,
-        data,
-        timestamp: Date.now(),
-      })
-      return response?.res ?? response
-    },
-    { type, data },
-  )
 }
 
 async function getAlarm(
